\name{geohippos-package}
\alias{geohippos-package}
\alias{geohippos}
\docType{package}
\title{
\packageTitle{geohippos}
}
\description{
\packageDescription{geohippos}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{geohippos}
\packageIndices{geohippos}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{geohippos}


}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
