% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohippos.r
\docType{data}
\name{geohippos}
\alias{geohippos}
\title{a geohippo}
\format{
character
}
\usage{
geohippos
}
\description{
To print the geohippo just use cat
}
\examples{
cat(geohippos)
}
\author{
Bernd Gruber
}
\keyword{datasets}
