% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.epos.r
\name{gl.epos}
\alias{gl.epos}
\title{Runs epos on a genlight/dartR obejct}
\usage{
gl.epos(
  x,
  epos.path,
  sfs = NULL,
  minbinsize = 1,
  folded = TRUE,
  l = NULL,
  u = 5e-09,
  boot = 0,
  upper = 0.975,
  lower = 0.025,
  other.options = "",
  plot.out = TRUE,
  verbose = NULL
)
}
\arguments{
\item{x}{dartR/genlight object}

\item{epos.path}{path to epos and other required programs (epos, epos2plot are always required and bootSfs in case a bootstrap and confidence estimate is required )}

\item{sfs}{if no sfs is provided function gl.sfs(x, minbinsize=1, singlepop=TRUE) is used to calculate the sfs that is provided to epos}

\item{minbinsize}{remove bins from the left of the sfs. if you run epos from a genlight object the sfs is calculated by the function (using gl.sfs) and as default minbinsize is set to 1 (the monomorphic loci of the sfs are removed). This parameter is ignored if sfs is provide via the sfs parameter (see below). Be aware even if you genlight object has more than one population the sfs is calculated with singlepop set to true (one sfs for all individuals) as epos does not work with multidimensional sfs)}

\item{folded}{if set to TRUE (default) a folded sfs (minor allele frequency sfs) is returned. If set to FALSE then an unfolded (derived allele frequency sfs) is returned. It is assumed that 0 is homozygote for the reference and 2 is homozygote for the derived allele. So you need to make sure your coding is correct. option -U in epos.}

\item{l}{length of sequences (including monomorphic and polymorphic sites). If the sfs is provided with minbinsize=1 (default) then l needs to be specified. 
option -l in epos}

\item{u}{mutation rate. If not provided the default value of epos is used (5e-9).
option -u in epos}

\item{boot}{if set to a value >0 the programm bootSfs is used to provide multiple
bootstrapped sfs, which allows to calculate confidence intervals of the historic Ne
sizes. Be aware the runtime can be extended. default:0 no bootstrapped simulations are
run, otherwise boot number of bootstraps are run (option -i in bootSfs)}

\item{upper}{upper quantile of the bootstrap (only used if boot>0). default 0.975.
(option -u in epos2plot)}

\item{lower}{lower quantile of the bootstrap (only used if boot>0). default 0.025.
(option -l in epos2plot)}

\item{other.options}{additional options for epos (e.g -m, -x etc.)}

\item{plot.out}{if set to true a timeseries plot of Ne (with bootstrap is
calculated)}

\item{verbose}{Verbosity: 0, silent or fatal errors; 1, begin and end; 2,
progress log ; 3, progress and results summary; 5, full report [default 2,
unless specified using gl.set.verbosity].}
}
\value{
returns a table of the Ne over time and a ggplot using estimate of times
and Ne (see epos.doc)
}
\description{
This is an implementation/wrapper to run epos \url{https://github.com/EvolBioInf/epos}. It relies on compiled version of epos, epos2plot and if a bootstrap output is required bootSfs. For more details on how to install these programs via \url{https://github.com/EvolBioInf/epos} and look out for the manual epos.pdf (in the doc folder of the repository \url{https://github.com/EvolBioInf/epos/blob/master/doc/epos.pdf}. The binaries need to be provided in a single folder. There are compiled versions available and also the necessary dlls for windows need to be provided (Under Linux/MacOS gls, blas need to be installed on your system)
}
\examples{
require(dartR.data)
gl.epos(possums.gl[1:60,])
gl.msfs(possums.gl[c(1:5,31:33),], minbinsize=1)
}
\references{
Lynch M, Haubold B, Pfaffelhuber P, Maruki T. Inference of Historical
Population-Size Changes with Allele-Frequency Data. G3 (Bethesda). 2020 Jan 7;10(1):211-223. doi: 10.1534/g3.119.400854. PMID: 31699776; PMCID: PMC6945023.
}
\author{
Custodian: Bernd Gruber
}
