% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.LinkNe.r
\name{gl.LinkNe}
\alias{gl.LinkNe}
\title{Runs LinkNe on a genlight/dartR obejct}
\usage{
gl.LinkNe(
  x,
  outfile = "LinkNe",
  L = 1e+08,
  LinkNe.path,
  perl = FALSE,
  temp.path = tempdir(),
  other.args = "",
  verbose = 1
)
}
\arguments{
\item{x}{genlight/dartR object}

\item{outfile}{name of the output file}

\item{L}{length of a single chromosome (used to create the map file)}

\item{LinkNe.path}{path to LinkNe (could be a binary (e.g. LinkNe.exe) or the perl version (LinkNe.pl))}

\item{perl}{switch to run the perl version instead of the binary. To run the perl version your system must have perl installed and accessible from any folder (e.g. in the path under windows)}

\item{other.args}{additional arguments passed to LinkNe (\link[LinkNe settings]{https://github.com/chollenbeck/LinkNe} )}

\item{verbosity}{(though we cannot suppress console output during LinkNe runs)}
}
\description{
This is an implementation/wrapper to run LinkNe \link[Hollenbeck et al. 2016]{http://www.nature.com/hdy/journal/v117/n4/abs/hdy201630a.html}.
}
\author{
Custodian: Bernd Gruber
}
