% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gl.stairway2.r
\name{gl.stairway2}
\alias{gl.stairway2}
\title{wrapper function to run stairway2}
\usage{
gl.stairway2(
  x,
  outfile = NULL,
  stairway.path,
  simfolder = NULL,
  minbinsize = 1,
  maxbinsize = NULL,
  pct_training = 0.67,
  nrand = NULL,
  stairway_plot_dir = "stairway_plot_es",
  nreps = 200,
  seed = NULL,
  mu = NULL,
  gentime = NULL,
  L = NULL,
  plot_title = "Ne",
  xmin = 0,
  xmax = 0,
  ymin = 0,
  ymax = 0,
  xspacing = 2,
  yspacing = 2,
  fontsize = 12,
  run = FALSE,
  parallel = 1,
  verbose = NULL,
  sfs = NULL,
  cleanup = TRUE
)
}
\description{
Runs using a dartR/genlight object
}
